Loader: Consolidate functions into a single file
LoadYAML now accepts an io.Reader
